--- api/src/instructors/dto/create.ts
@@ -1,29 +1,0 @@
-export class CreateInstructorDTO {
-  firstName: string;
-
-  lastName: string;
-
-  phone: string;
-
-  description: string;
-
-  email: string;
-
-  password: string;
-
-  image_url: string;
-}
-
-export class InstructorPostBody {
-  firstName: string;
-
-  lastName: string;
-
-  phone: string;
-
-  description: string;
-
-  email: string;
-
-  password: string;
-}
