+++ api/src/instructors/instructors.controller.ts
@@ -7,7 +7,7 @@ import {
   UploadedFile,
   UseInterceptors,
 } from '@nestjs/common';
-import { InstructorPostBody } from './dto/create';
+import { InstructorPostBody } from './dto/create.dto';
 import { FileInterceptor } from '@nestjs/platform-express';
 import { InstructorsService } from './instructors.service';
